Добавлено разделение ответственности на "Лексер" и "Парсер". Отрефакторена реализция фильтрации тегов и токенов
